refactor: v.0.1.2 improve notes grid rendering
before : It renders each note and calculates grid height, therefore total rendering count is twice the number of
the notes.
after : It renders all notes and calculates grid heights, then re-renders
all notes, therefore the total rendering count is just two. But, the first
rendering is opacity 0. So, only the second rendering is shown to the user